requirements.txt
# Django and Web Framework
Django==5.0.1
djangorestframework==3.14.0

# Environment Variables
python-decouple==3.8

# CLI Enhancement
rich==13.7.0

# API Integration
requests==2.31.0

# Date & Time
python-dateutil==2.8.2

# Note: pandas and numpy require pre-built wheels on Windows
# Install separately if needed: pip install pandas numpy
